Improve the KPI optimistic updates
* use the new title immediately
* extract the logic into a hook and move it to its own file

JIRA: RAIL-4488
