+++ libs/sdk-ui-dashboard/src/presentation/widget/kpi/EditModeDashboardKpi/EditModeDashboardKpi.tsx
@@ -1,9 +1,9 @@
 // (C) 2022 GoodData Corporation
-import React, { useCallback, useEffect, useMemo, useState } from "react";
+import React, { useCallback, useEffect, useMemo } from "react";
 import cx from "classnames";
 import { useIntl } from "react-intl";
 import noop from "lodash/noop";
-import { areObjRefsEqual, widgetRef } from "@gooddata/sdk-model";
+import { widgetRef } from "@gooddata/sdk-model";
 import { useBackendStrict, useWorkspaceStrict } from "@gooddata/sdk-ui";
 
 import {
@@ -18,11 +18,6 @@ import {
     uiActions,
     useWidgetSelection,
     selectIsDashboardSaving,
-    useDashboardEventsContext,
-    isDashboardCommandStarted,
-    isDashboardKpiWidgetMeasureChanged,
-    DashboardEventHandler,
-    isDashboardCommandFailed,
 } from "../../../../model";
 import { DashboardItemHeadline, DashboardItemKpi } from "../../../presentationComponents";
 import { useDashboardComponentsContext } from "../../../dashboardContexts";
@@ -30,6 +25,7 @@ import { useDashboardComponentsContext } from "../../../dashboardContexts";
 import { ConfigurationBubble } from "../../common";
 import { getKpiResult, KpiRenderer, useKpiData, useKpiExecutionDataView } from "../common";
 import { IDashboardKpiProps } from "../types";
+import { useOptimisticMeasureUpdate } from "./useOptimisticMeasureUpdate";
 
 export const EditModeDashboardKpi = (props: IDashboardKpiProps) => {
     const {
@@ -52,41 +48,7 @@ export const EditModeDashboardKpi = (props: IDashboardKpiProps) => {
 
     const KpiConfigurationComponent = KpiWidgetComponentSet.configuration.WidgetConfigPanelComponent;
 
-    // handle loading indicator when changing measure
-    const [isChangingMeasure, setIsChangingMeasure] = useState(false);
-    const { registerHandler, unregisterHandler } = useDashboardEventsContext();
-    useEffect(() => {
-        const onMeasureChangingStarted: DashboardEventHandler = {
-            eval: (evt: any) => {
-                return (
-                    isDashboardCommandStarted(evt) &&
-                    evt.payload.command.type === "GDC.DASH/CMD.KPI_WIDGET.CHANGE_MEASURE" &&
-                    areObjRefsEqual(evt.payload.command.payload.ref, kpiWidget.ref)
-                );
-            },
-            handler: () => setIsChangingMeasure(true),
-        };
-        const onMeasureChangingEnded: DashboardEventHandler = {
-            eval: (evt: any) => {
-                return (
-                    (isDashboardKpiWidgetMeasureChanged(evt) &&
-                        areObjRefsEqual(evt.payload.ref, kpiWidget.ref)) ||
-                    (isDashboardCommandFailed(evt) &&
-                        evt.payload.command.type === "GDC.DASH/CMD.KPI_WIDGET.CHANGE_MEASURE" &&
-                        areObjRefsEqual(evt.payload.command.payload.ref, kpiWidget.ref))
-                );
-            },
-            handler: () => setIsChangingMeasure(false),
-        };
-
-        registerHandler(onMeasureChangingStarted);
-        registerHandler(onMeasureChangingEnded);
-
-        return () => {
-            unregisterHandler(onMeasureChangingStarted);
-            unregisterHandler(onMeasureChangingEnded);
-        };
-    }, [kpiWidget.ref, registerHandler, unregisterHandler]);
+    const { isChangingMeasure, titleToShow } = useOptimisticMeasureUpdate(kpiWidget);
 
     const backend = useBackendStrict(customBackend);
     const workspace = useWorkspaceStrict(customWorkspace);
@@ -181,10 +143,7 @@ export const EditModeDashboardKpi = (props: IDashboardKpiProps) => {
                 return null;
             }}
             renderHeadline={(clientHeight) => (
-                <DashboardItemHeadline
-                    title={isChangingMeasure ? "" : kpiWidget.title}
-                    clientHeight={clientHeight}
-                />
+                <DashboardItemHeadline title={titleToShow} clientHeight={clientHeight} />
             )}
             isSelectable={isSelectable}
             isSelected={isSelected}
